Add support for weekly repeating events
This adds support for events that repeats on chosen days of the week. E.g.
repeating every Tuesday and Monday.

The solution isn't the most elegant, as we parse the date again from a
previously parsed, then formatted source (the `date` variable), but for a more
elegant solution we would need to refactor other methods as well.

For the repeating the following format is used (note the dates beneth):
```
* Pretend to do some sport
  :PROPERTIES:
  :ID:        a3f62a3f-5d6f-4956-9adc-a94b73438b34
  :LOCATION:  Local Gym
  :STATUS:    CONFIRMED
  :ATTENDING: ATTENDING
  :ATTENDEES:
  :END:
  :LOGBOOK:
  CLOCK: [2023-07-15 Sat 10:15]--[2023-07-15 Sat 10:30] =>  0:00
  :END
<2023-07-17 Mon 10:15-10:30 +1w>
<2023-07-18 Tue 10:15-10:30 +1w>
<2023-07-19 Wed 10:15-10:30 +1w>
<2023-07-20 Thu 10:15-10:30 +1w>
```

Other solution could be to use a lisp expression to setup, but for me that is
too obsucre, hard to read at glance. E.g.:
```
** Daily meeting
   <%%(memq (calendar-day-of-week date) '(1 2 3 4 5))>
```
source: https://orgmode.org/worg/org-faq.html#orga9d21b4